requirements.txt


flask==3.0.2
python-dotenv==1.0.0
PyPDF2==3.0.1
langchain-anthropic==0.1.0
anthropic==0.20.0